Added new methods that aid in Fish movement.
Added "addFish" method, which places a fish on the board and tells the Fish object where it exists within the environment.

Added "isValidAdress", which is a boolean method that determines whether a specific x and y address exist within the environment.

Added "moveFish", which uses "isValidAddress" and "swapNodes" methods to move the fish within the environment and updates the Fish object's x and y attributes.  Accepts a Fish object and an integer array containing the x and y coordinates to move to.