Modified movement methods
The methods "wander" and "move" work the same in principle, but instead of changing the value of the Fish object's attributes x and y, they return new x and y coordinates with the appropriate math assigned to them.  These coordiates are fed to the environment to determine if these coordinates are valid or not.