Swapped X and Y values in Environment methods
When editting the environmentGrid matrix array, although conterintuitive, all entries should be denoted as: environmentGrid[Y][X].  This is becacuse of how the board is drawn.  Technically, you can still write your input as environmentGrid[X][Y], but the result may not end up where you want it to be.